Enable honeypot reporting and editing for a job
