+++ cvat/apps/engine/rules/jobs.rego
@@ -253,3 +253,18 @@ allow if {
     organizations.has_perm(organizations.WORKER)
     is_task_staff
 }
+
+allow if {
+    input.scope in {utils.VIEW_HONEYPOTS, utils.UPDATE_HONEYPOTS}
+    utils.has_perm(utils.USER)
+    utils.is_sandbox
+    is_task_staff
+}
+
+allow if {
+    input.scope in {utils.VIEW_HONEYPOTS, utils.UPDATE_HONEYPOTS}
+    input.auth.organization.id == input.resource.organization.id
+    organizations.has_perm(organizations.SUPERVISOR)
+    utils.has_perm(utils.USER)
+    is_task_staff
+}
